Refactor code for readability in AuditSecurityConfig
The code to fetch the current user was moved to a separate private method called "getCurrentUser". This was done to make the auditorAware method cleaner and easier to understand. It also makes the code more reusable and organized. Improvements were made to the class and method comments to better describe their functions.