Refactor User class and add object comparison methods
The User.java class was refactored for the better reading experience. Import statements were formatted for easy reading and references. Locale.ENGLISH was added to the toLowerCase method to ensure uniformity across different Server setups. Equals and hashCode methods were overridden to enable object comparison based on email, which is a unique field. The serialVersionUID was also updated due to the structural change in the class. Also, "@Valid" is added to the Set<Reservation> to ensure the persistence of valid reservations only.